factor out three copies of ulp_distance
