+++ tests/numberparsingcheck.cpp
@@ -4,7 +4,7 @@
 #include <cstdio>
 #include <cstdlib>
 
-
+#include "number_comparison.h"
 
 #ifndef JSON_TEST_NUMBERS
 #define JSON_TEST_NUMBERS
@@ -46,31 +46,7 @@ void found_unsigned_integer(uint64_t result, const uint8_t *buf);
 
 #endif
 
-// ulp distance
-// Marc B. Reynolds, 2016-2019
-// Public Domain under http://unlicense.org, see link for details.
-// adapted by D. Lemire
-inline uint32_t f32_ulp_dist(float a, float b) {
-  uint32_t ua, ub;
-  memcpy(&ua, &a, sizeof(ua));
-  memcpy(&ub, &b, sizeof(ub));
-  if ((int32_t)(ub ^ ua) >= 0)
-    return (int32_t)(ua - ub) >= 0 ? (ua - ub) : (ub - ua);
-  return ua + ub + 0x80000000;
-}
 
-// ulp distance
-// Marc B. Reynolds, 2016-2019
-// Public Domain under http://unlicense.org, see link for details.
-// adapted by D. Lemire
-inline uint64_t f64_ulp_dist(double a, double b) {
-  uint64_t ua, ub;
-  memcpy(&ua, &a, sizeof(ua));
-  memcpy(&ub, &b, sizeof(ub));
-  if ((int64_t)(ub ^ ua) >= 0)
-    return (int64_t)(ua - ub) >= 0 ? (ua - ub) : (ub - ua);
-  return ua + ub + 0x80000000;
-}
 
 int parse_error;
 char *fullpath;
@@ -187,7 +163,7 @@ void found_float(double result, const uint8_t *buf) {
     fprintf(stderr, "       %.32s whereas strtod gives\n", buf);
     fprintf(stderr, "       %.128e,", expected);
     fprintf(stderr, " while parsing %s \n", fullpath);
-    fprintf(stderr, " ===========  ULP:  %u,", (unsigned int)f64_ulp_dist(expected, result));
+    fprintf(stderr, " ===========  ULP:  %u,", (unsigned int)number_tests::f64_ulp_dist(expected, result));
     parse_error |= PARSE_ERROR;
   }
 }
